refactor(runner): move err/test runners to core
- relocate run_err_cmd/run_test_cmd and filters to core::runner
- cmds/rust/runner keeps shell wrappers delegating to core
- bun/deno call core::runner, no reach into cmds/rust
